+++ app/src/main/java/com/xsas/appsaludtest/dominio/presentadores/PresentadorListaEstadoVivienda.kt
@@ -23,7 +23,7 @@ class PresentadorListaEstadoVivienda : Presentador<List<EstadoVivienda>>(){
             assert(jsonObject != null)
 
             val estadoVivienda = EstadoVivienda(
-                json.optInt("id_estado") as BigInteger,
+                json.optInt("id_estado"),
                 json.optString("estado_vivienda")
             )
